Make proper use of the unblockable right
This commit rewrites the internals to make use of the
unblockable right. This replaces the check for the staff group.
To ensure maximum compatiblity with the previous versions, two
new global variables are introduced:
wgStaffPowersStewardGroupName with the default 'steward'
and wgStaffPowersShoutWikiMessages with the default true.
wgStaffPowersStewardGroupName contains the groupname of the
steward group, a group only blockable by those who have the
unblockable right. If this global is set to false, NULL or an
empty string, it will be ignored.
wgStaffPowersShoutWikiMessages toggles the ShoutWiki specific
messages when blocking either stewards or those with the
unblockable right. Its functionality has also been fixed, as it
previously didn't work. In addition to these changes, the Dutch
translation has also been updated.
Edit: Fixed naming, per PHP naming conventions
Edit 2: Applied fixes advised by Jack Phoenix. I also forgot to
mention that the this patchset also introduces the shorthand
array syntax.

Change-Id: I1fccec3b74dd9b0b98994b81e3a3426efca898c0